cfg: tweak new file handling
Ensure new files to be saved are stored as real files and not symlinks.

We used to replace all files in the RAM disk with symlinks to
$MOUNTPOUNT/<file> for space / version considerations, and then let cfg
load overwrite the files we have stored in cfg. This means that if new
files referred in SAVE are added, then those will be represented as
symlinks (making editing tricky) and will be SAVED as symlinks as well.

Fix it so now files listed in SAVE will be copied rather than symlinked,
so cfg commit does the right thing.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>